4th Commit - Improved Performance
Worked out some of th eproblems with the obstacle avoidance code. Car was having problems because the turnspeed for linetracking and for obstacle avoidance have to be different. Additionally modified the avoidance behaviour so that the car double checks whether the obstacle is still in its way. If the obstacle (a person) moved away then the bot will continue on its linetack,