11th Commit - Added behavior Commands
Car will now remain motionless unitll the 'G' key (go) is pressed. Car can be put in a halted state using the 'S' key (stop). Car testing is not intrisically liked to using the raspberry pi as the python script must be running to register kestrokes which undo halted state.