14th Commit - Code Almost Working
Code can do everything except merge back onto the black line. To try and fix this some untested behavior was added where the car will run slower untill a mergetick counter is satisfied. the hope is that by this point the car will have converged onto the line. ALSO added some cola selection buttons to the python script. if one presses 'P' car with search for sprite. As the default or when 'O' is pressed the car will search for Coca-Cola.